test(workflows): add tests for rolling-release manual dispatch workflow
- Added a new test module to validate the manual dispatch workflow contracts.
- Ensure the manual dispatch exposes the `force_dependency_binary_rebuild` boolean input.
- Validate that dependency manifest change step respects the manual rebuild input.
- Refactored workflow test helpers into shared support module for reuse.
- Removed duplicated test helper functions from main test module.
