+++ tests/workflows/test_rolling_release_workflow.py
@@ -24,13 +24,17 @@ from __future__ import annotations
 import os
 import re
 import shlex
-from collections.abc import Mapping
 from pathlib import Path
-from typing import Any, cast
 
 import pytest
-from ruamel.yaml import YAML
 
+from tests.workflows.rolling_release_workflow_test_support import (
+    _find_step_by_name,
+    _get_job_dict,
+    _get_needs_list,
+    _github_expression_value,
+    _load_workflow_mapping,
+)
 from tests.workflows.workflow_test_helpers import (
     install_components_script,
     lint_crates_from_resolution_constants,
@@ -161,83 +165,7 @@ def test_install_components_uses_only_matrix_target_rustc_dev(workflow_text: str
     assert not rust_src_commands, (
         "install step must not install rust-src because it conflicts with "
         "targeted rustc-dev payloads on some runners"
-    )
-
-
-def _load_workflow_mapping(yaml_text: str) -> dict[str, object]:
-    """Load YAML text and return workflow mapping."""
-    parsed = YAML(typ="safe").load(yaml_text)
-    match parsed:
-        case dict() as workflow_mapping:
-            return workflow_mapping
-        case _:
-            pytest.fail("rolling-release workflow must parse to a mapping")
-
-
-def _get_job_dict(jobs: Mapping[str, Any], job_name: str) -> dict[str, Any]:
-    """Return the requested job mapping from the jobs map."""
-    match jobs.get(job_name):
-        case dict() as job_dict:
-            return job_dict
-        case _:
-            if job_name == "jobs":
-                pytest.fail("rolling-release workflow must declare jobs")
-            pytest.fail(f"rolling-release workflow must declare {job_name} job")
-
-
-def _workflow_dispatch_inputs(workflow_mapping: Mapping[str, Any]) -> dict[str, Any]:
-    """Return the `workflow_dispatch.inputs` mapping."""
-    on_mapping = workflow_mapping.get("on")
-    match on_mapping:
-        case {"workflow_dispatch": {"inputs": dict() as inputs}}:
-            return inputs
-        case {"workflow_dispatch": dict()}:
-            pytest.fail("workflow_dispatch must declare an inputs mapping")
-        case {"workflow_dispatch": _}:
-            pytest.fail("workflow_dispatch trigger must be a mapping")
-        case _:
-            pytest.fail("rolling-release workflow must declare workflow_dispatch")
-
-
-def _github_expression_value(value: object) -> str:
-    """Return a GitHub Actions expression with wrapper delimiters removed."""
-    if not isinstance(value, str):
-        pytest.fail("workflow expression value must be a string")
-    stripped = value.strip()
-    if stripped.startswith("${{") and stripped.endswith("}}"):
-        return stripped[3:-2].strip()
-    return stripped
-
-
-def _get_needs_list(publish_job: dict[str, Any]) -> list[str]:
-    """Return publish job dependency names as a list."""
-    needs: str | list[str] | None = publish_job.get("needs")
-    match needs:
-        case str():
-            return [needs]
-        case list():
-            if all(isinstance(item, str) for item in needs):
-                return cast(list[str], needs)
-            pytest.fail("publish job needs list must contain only strings")
-        case _:
-            pytest.fail("publish job needs must be a string or list")
-
-
-def _find_step_by_name(steps: object, name: str) -> dict[str, object] | None:
-    """Find a step dict by its name."""
-    match steps:
-        case list():
-            pass
-        case _:
-            pytest.fail("publish job must declare steps")
-
-    for step in steps:
-        match step:
-            case {"name": step_name} if step_name == name:
-                return step
-            case _:
-                continue
-    return None
+)
 
 
 def test_publish_job_runs_even_if_build_lints_fails(workflow_text: str) -> None:
@@ -259,100 +187,6 @@ def test_publish_job_runs_even_if_build_lints_fails(workflow_text: str) -> None:
         "download step must continue on error so zero-artefact runs can fall "
         "through to has_assets=false"
     )
-
-
-def test_manual_dispatch_exposes_force_dependency_binary_rebuild_input(
-    workflow_text: str,
-) -> None:
-    """Ensure manual dispatch exposes an explicit dependency rebuild switch."""
-    workflow_mapping = _load_workflow_mapping(workflow_text)
-    inputs = _workflow_dispatch_inputs(workflow_mapping)
-
-    input_mapping = inputs.get("force_dependency_binary_rebuild")
-    assert isinstance(input_mapping, dict), (
-        "workflow_dispatch must declare force_dependency_binary_rebuild input"
-    )
-    assert input_mapping.get("type") == "boolean", (
-        "force_dependency_binary_rebuild must be a boolean workflow input"
-    )
-    assert input_mapping.get("required") is False, (
-        "force_dependency_binary_rebuild must remain optional for manual "
-        "rolling-release republishes"
-    )
-    assert input_mapping.get("default") is False, (
-        "force_dependency_binary_rebuild must default to false so manual runs "
-        "do not rebuild dependency binaries implicitly"
-    )
-
-    description = input_mapping.get("description")
-    assert isinstance(description, str), (
-        "force_dependency_binary_rebuild must document its recovery purpose"
-    )
-    assert description.strip(), (
-        "force_dependency_binary_rebuild must document its recovery purpose"
-    )
-    description_lower = description.lower()
-    assert "dependency" in description_lower, (
-        "force_dependency_binary_rebuild description must explain that it "
-        "forces a dependency binary rebuild"
-    )
-    assert "rebuild" in description_lower, (
-        "force_dependency_binary_rebuild description must explain that it "
-        "forces a dependency binary rebuild"
-    )
-
-
-def test_dependency_manifest_change_step_only_forces_manual_rebuild_when_requested(
-    workflow_text: str,
-) -> None:
-    """Ensure manual dispatch force input gates dependency rebuilds."""
-    workflow_mapping = _load_workflow_mapping(workflow_text)
-    jobs = _get_job_dict(workflow_mapping, "jobs")
-    change_job = _get_job_dict(jobs, "dependency-manifest-changes")
-    check_step = _find_step_by_name(
-        change_job.get("steps"),
-        "Check whether dependency manifest changed",
-    )
-    assert check_step is not None, (
-        "dependency-manifest-changes job must check whether the dependency "
-        "manifest changed"
-    )
-    run_script = check_step.get("run", "")
-    assert isinstance(run_script, str), "change-detection step must have a run script"
-
-    assert "github.event.inputs.force_dependency_binary_rebuild" in run_script, (
-        "change-detection step must read the manual "
-        "force_dependency_binary_rebuild input"
-    )
-    assert (
-        '[[ "${{ github.event_name }}" == "workflow_dispatch" ]]' in run_script
-    ), "change-detection step must branch explicitly on workflow_dispatch"
-    assert (
-        '[[ "${{ github.event.inputs.force_dependency_binary_rebuild }}" == "true" ]]'
-        in run_script
-    ), (
-        "workflow_dispatch path must only set should_build=true when the "
-        "manual force input is true"
-    )
-    assert re.search(
-        r'if\s+\[\[\s+"\$\{\{\s+github\.event_name\s+\}\}"\s+==\s+"workflow_dispatch"\s+\]\]'
-        r'\s*;\s*then\s*echo\s+"should_build=true"\s+>>\s+"\$GITHUB_OUTPUT"',
-        run_script,
-        re.DOTALL,
-    ) is None, (
-        "workflow_dispatch must not unconditionally rebuild dependency "
-        "binaries on every manual run"
-    )
-    assert "echo \"should_build=false\" >> \"$GITHUB_OUTPUT\"" in run_script, (
-        "manual runs without force input must leave should_build=false"
-    )
-    assert "git diff --quiet" in run_script, (
-        "push-based dependency manifest diff detection must remain in place"
-    )
-    assert "installer/dependency-binaries.toml" in run_script, (
-        "push-based dependency manifest diff detection must remain in place"
-    )
-
 
 def test_restore_step_guards_against_missing_dependency_assets(
     workflow_text: str,
